Moving sub atlas cutting to separate class
